ATMcash v.1 (NXT based)
This is the first commit for the NXT-based ATMcash project.

This is the beginning of the project that is started in this fashion.

AT a later date that will soon be announced, a completely new backend with many improvements and changes will be implemented. Along with more information regarding the swap from old to new chain. Check back for new updates. Thanks!